plugins/lcr_mpi_tool/requirements.txt
# ── Deep Learning ──────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0

# ── Numerical & Scientific Computing ───────────────────────────
numpy>=1.24.0
scipy>=1.10.0

# ── HDF5 / .mat File I/O ───────────────────────────────────────
h5py>=3.8.0

# ── Image Processing ───────────────────────────────────────────
opencv-python>=4.7.0
Pillow>=9.5.0

# ── Visualization ──────────────────────────────────────────────
matplotlib>=3.7.0

# ── Tensor Utilities ───────────────────────────────────────────
einops>=0.6.1

# ── Progress Bar ───────────────────────────────────────────────
tqdm>=4.65.0
